+++ webapp/src/main/java/com/infoshare/fourfan/repository/AdminRepository.java
@@ -15,5 +15,5 @@ public interface AdminRepository {
 
     Product findProductById(Integer id) throws IOException;
 
-    void editProduct(Long id, Product product) throws IOException;
+    void editProduct(Integer id, Product product) throws IOException;
 }
